chore: refactor views to protected
